Bump versions
As this repo hadn't been looked at for a while, some libraries were in need of an update. This brings the Ruby and Rails versions up to latest, and drags along a number of other libraries as well.

Simultaneously, I'll set up automated alerts on the repo to ensure this doesn't have to be done in such a big bang way next time.

Signed-off-by: jonathankerr <jonathan.kerr@education.gov.uk>